refactor: move RelOutputTest to SubstraitRelNodeConverterTest
